+++ src/utils/index.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = {
+  ...require('./createGossipRpc'),
+  ...require('./shuffle')
+}
